+++ packages/ipfs-proxy/.prettierrc
@@ -1,0 +1,6 @@
+{
+	"printWidth": 140,
+	"singleQuote": true,
+	"semi": true,
+	"useTabs": true
+}
