+++ packages/ipfs-proxy/vitest.config.mts
@@ -1,0 +1,11 @@
+import { defineWorkersConfig } from '@cloudflare/vitest-pool-workers/config';
+
+export default defineWorkersConfig({
+	test: {
+		poolOptions: {
+			workers: {
+				wrangler: { configPath: './wrangler.json' },
+			},
+		},
+	},
+});
