+++ packages/ipfs-proxy/worker-configuration.d.ts
@@ -1,0 +1,4 @@
+// Generated by Wrangler
+// After adding bindings to `wrangler.json`, regenerate this interface via `npm run cf-typegen`
+interface Env {
+}
